+++ lang/test/org/partiql/lang/eval/EvaluatingCompilerOrderByTests.kt
@@ -109,39 +109,47 @@ class EvaluatingCompilerOrderByTests : EvaluatorTestBase() {
                 "[{'productId': 4, 'supplierId_nulls': NULL}, {'productId': 5, 'supplierId_nulls': NULL}, {'productId': 8, 'supplierId_nulls': NULL}, {'productId': 9, 'supplierId_nulls': NULL}, {'productId': 10, 'supplierId_nulls': NULL}, {'productId': 6, 'supplierId_nulls': 11}, {'productId': 7, 'supplierId_nulls': 11}, {'productId': 1, 'supplierId_nulls': 10}, {'productId': 2, 'supplierId_nulls': 10}, {'productId': 3, 'supplierId_nulls': 10}]"
             ),
             // should group and order by asc sellerId
+            // @TODO: Planner does NOT support GROUP BY yet. Need to add support for the following 7 tests
             EvaluatorTestCase(
                 "SELECT sellerId FROM orders GROUP BY sellerId ORDER BY sellerId ASC",
-                "[{'sellerId': 1}, {'sellerId': 2}]"
+                "[{'sellerId': 1}, {'sellerId': 2}]",
+                target = EvaluatorTestTarget.COMPILER_PIPELINE
             ),
             // should group and order by desc sellerId
             EvaluatorTestCase(
                 "SELECT sellerId FROM orders GROUP BY sellerId ORDER BY sellerId DESC",
-                "[{'sellerId': 2}, {'sellerId': 1}]"
+                "[{'sellerId': 2}, {'sellerId': 1}]",
+                target = EvaluatorTestTarget.COMPILER_PIPELINE
             ),
             // should group and order by DESC (NULLS FIRST as default)
             EvaluatorTestCase(
                 "SELECT supplierId_nulls FROM products_sparse GROUP BY supplierId_nulls ORDER BY supplierId_nulls DESC",
-                " [{'supplierId_nulls': NULL}, {'supplierId_nulls': 11}, {'supplierId_nulls': 10}]"
+                " [{'supplierId_nulls': NULL}, {'supplierId_nulls': 11}, {'supplierId_nulls': 10}]",
+                target = EvaluatorTestTarget.COMPILER_PIPELINE
             ),
             // should group and order by ASC (NULLS LAST as default)
             EvaluatorTestCase(
                 "SELECT supplierId_nulls FROM products_sparse GROUP BY supplierId_nulls ORDER BY supplierId_nulls ASC",
-                "[{'supplierId_nulls': 10}, {'supplierId_nulls': 11}, {'supplierId_nulls': NULL}]"
+                "[{'supplierId_nulls': 10}, {'supplierId_nulls': 11}, {'supplierId_nulls': NULL}]",
+                target = EvaluatorTestTarget.COMPILER_PIPELINE
             ),
             // should group and place nulls first (asc as default)
             EvaluatorTestCase(
                 "SELECT supplierId_nulls FROM products_sparse GROUP BY supplierId_nulls ORDER BY supplierId_nulls NULLS FIRST",
-                "[{'supplierId_nulls': NULL}, {'supplierId_nulls': 10}, {'supplierId_nulls': 11}]"
+                "[{'supplierId_nulls': NULL}, {'supplierId_nulls': 10}, {'supplierId_nulls': 11}]",
+                target = EvaluatorTestTarget.COMPILER_PIPELINE
             ),
             // should group and place nulls last (asc as default)
             EvaluatorTestCase(
                 "SELECT supplierId_nulls FROM products_sparse GROUP BY supplierId_nulls ORDER BY supplierId_nulls NULLS LAST",
-                "[{'supplierId_nulls': 10}, {'supplierId_nulls': 11}, {'supplierId_nulls': NULL}]"
+                "[{'supplierId_nulls': 10}, {'supplierId_nulls': 11}, {'supplierId_nulls': NULL}]",
+                target = EvaluatorTestTarget.COMPILER_PIPELINE
             ),
             // should group and order by asc and place nulls first
             EvaluatorTestCase(
                 "SELECT supplierId_nulls FROM products_sparse GROUP BY supplierId_nulls ORDER BY supplierId_nulls ASC NULLS FIRST",
-                "[{'supplierId_nulls': NULL}, {'supplierId_nulls': 10}, {'supplierId_nulls': 11}]"
+                "[{'supplierId_nulls': NULL}, {'supplierId_nulls': 10}, {'supplierId_nulls': 11}]",
+                target = EvaluatorTestTarget.COMPILER_PIPELINE
             ),
 
             // DIFFERENT DATA TYPES
@@ -289,8 +297,7 @@ class EvaluatingCompilerOrderByTests : EvaluatorTestBase() {
     @ArgumentsSource(ArgsProviderValid::class)
     fun validTests(tc: EvaluatorTestCase) = runEvaluatorTestCase(
         tc = tc.copy(
-            excludeLegacySerializerAssertions = true,
-            targetPipeline = EvaluatorTestTarget.COMPILER_PIPELINE, // planner & phys. alg. have no support for ORDER BY (yet)
+            excludeLegacySerializerAssertions = true
         ),
         session = session
     )